fling-api/pyproject.toml
[tool.poetry]
name = "fling-api"
version = "0.1.3"
description = "Side Project Management API for Data Storage and Name Searches"
authors = ["Joshua McKenty <jmckenty@gmail.com>", "Anouk Ruhaak <anoukruhaak@gmail.com>"]
readme = "README.md"
packages = [{include = "api"}]

[tool.poetry.dependencies]
openai = "^0.27.4"
python-dotenv = "^1.0.0"
fastapi = "^0.95.0"
uvicorn = "^0.21.1"
boto3 = "^1.26.108"
python = "^3.9"
fling-core = {path="../fling-core", develop=true}
# fling-core = "^0.1.3"
cachetools = "^5.3.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"